[fix/RF-34] fix button component and stories
